move network stuff in netx package
